Some modifications
